Update SlideForm component to send the user to '/menu' path when the form is submitted #4
