+++ server/public/styles.css
@@ -2,6 +2,11 @@
   content: none;
 }
 
+.flex-wrap {
+  display: flex;
+  flex-wrap: wrap;
+}
+
 .justify-center {
   display: flex;
   justify-content: center;
@@ -147,4 +152,20 @@
   object-fit: cover;
   width: 300px;
   height: 300px;
+}
+
+.slides-overlay {
+  background-color: rgb(0 0 0 / 75%);
+  width: 100%;
+  height: 100vh;
+}
+
+.progress-bar-container {
+  position: absolute;
+  top: 8px;
+  width: 100%;
+}
+
+.progress-bar.active {
+  background-color: white;
 }
